feat : 페이지 명 수정 (childInfo -> onboarding), 기존 onboarding 페이지 제거, 루트 변경
