+++ src/routes.tsx
@@ -6,7 +6,6 @@ const HomePage = lazy(() => import('./pages/HomePage'));
 const LandingPage = lazy(() => import('./pages/LandingPage'));
 const MyPage = lazy(() => import('./pages/MyPage'));
 const LoginPage = lazy(() => import('./pages/LoginPage'));
-const IntroPage = lazy(() => import('./pages/IntroPage'));
 const OnboardingPage = lazy(() => import('./pages/OnboardingPage'));
 const ProblemHistoryPage = lazy(() => import('./pages/ProblemHistoryPage'));
 const ProblemUploadPage = lazy(() => import('./pages/ProblemUploadPage'));
@@ -49,17 +48,8 @@ const router = createBrowserRouter([
       },
       {
         path: '/onboarding',
-        element: (
-          <RequireAuth>
-            <OnboardingPage />
-          </RequireAuth>
-        ),
+        element: <OnboardingPage />,
       },
-      {
-        path: '/childIntro',
-        element: <IntroPage />,
-      },
-
       {
         path: '/login',
         element: (
